internal/grpc: move grpc API server code to its own package
Move grpc api server code to its own package.
